Move getResourceIrisFromStandoffTags to StandoffStuff
